+++ src/actions/clients/delete-client.ts
@@ -3,26 +3,41 @@
 import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
 
-export const deleteClientWithId = async(clientId:number) => {
+export const deleteClientWithId = async (clientId: number) => {
     try {
 
+
+        // Verificar que no este en un pedido
+        const relatedPedido = await prisma.pedido.findMany({
+            where: {
+                clienteId: clientId
+            }
+        });
+
+        if (relatedPedido.length > 0) {
+            return {
+                ok: false,
+                message: 'El cliente no se puede eliminar porque se encuentra en un pedido'
+            }
+        }
+
         await prisma.cliente.delete({
-            where : {
-                id:clientId
+            where: {
+                id: clientId
             }
         });
 
         revalidatePath('/clientes');
         return {
-            ok:true,
-            message:'Cliente eliminado!'
+            ok: true,
+            message: 'Cliente eliminado!'
         }
-        
+
     } catch (error) {
         console.error(error);
         return {
-            ok:false,
-            message:'Error al eliminar el cliente'
+            ok: false,
+            message: 'Error al eliminar el cliente'
         }
     }
 }
